refactor: separating interfaces into more specific files
