+++ src/domains/tasks/interfaces/ICreateTask.ts
@@ -1,6 +1,12 @@
+import { ITaskId } from "./ITaskRepository";
 import tasksStatus from "./taskStatus";
 
 export interface ITaskCreateParams {
   content: string;
   status: tasksStatus;
+}
+
+export interface ICreateTask extends ITaskId {
+  content: string;
+  status: tasksStatus
 }
